worker/requirements.txt
celery
redis
insightface
onnxruntime
opencv-python
ffmpeg-python
numpy
python-dotenv
requests
sqlalchemy
psycopg2-binary
